+++ deployment/terraform/variables.tf
@@ -135,6 +135,10 @@ variable "batch_ecs_image_minimum_cleanup_age" {
   default = "30m"
 }
 
+variable "tilegarden_root" {
+  default = ""
+}
+
 # Django
 variable "django_env" {}
 
@@ -156,4 +160,10 @@ variable "aws_cloudwatch_logs_policy_arn" {
 
 variable "pfb_app_alb_ingress_cidr_block" {
   type = "list"
+}
+
+# CloudFront distribution
+variable "tilegarden_api_gateway_domain_name" {}
+variable "cloudfront_price_class" {
+  default = "PriceClass_100"
 }