+++ src/django/pfb_analysis/models.py
@@ -11,7 +11,7 @@ import uuid
 import zipfile
 
 from django.conf import settings
-from django.contrib.gis.db.models import MultiPolygonField, PointField
+from django.contrib.gis.db.models import LineStringField, MultiPolygonField, PointField
 from django.contrib.gis.geos import GEOSGeometry, MultiPolygon, Polygon
 from django.contrib.postgres.fields import JSONField
 from django.core.files import File
@@ -20,6 +20,7 @@ from django.utils.text import slugify
 
 import botocore
 import boto3
+from django_countries.fields import CountryField
 import fiona
 from fiona.crs import from_epsg
 from localflavor.us.models import USStateField
@@ -40,11 +41,22 @@ SIMPLIFICATION_TOLERANCE_LESS = 0.0001
 SIMPLIFICATION_MIN_VALID_AREA_RATIO = 0.95
 
 
-def get_neighborhood_file_upload_path(instance, filename):
-    """ Upload each boundary file to its own directory """
-    return 'neighborhood_boundaries/{0}/{1}/{2}{3}'.format(slugify(instance.organization.name),
-                                                           instance.state_abbrev,
-                                                           instance.name,
+def get_neighborhood_file_upload_path(obj, filename):
+    """Upload each boundary file to its own directory
+
+    Upload file path should be unique for the organization.
+
+    Maintain backwards compatibility for previously-uploaded bounds with only state and no country
+    by not adding country to file path, but instead supporting paths to bounds outside the US
+    by using the three-letter country abbreviation instead of two-letter state abbreviation
+    in the file path.
+
+    Use three-letter country abbreviations instead of deafult two to avoid any conflicts with
+    two-letter state abbreviations (i.e., CA might be Canada or California.)
+    """
+    return 'neighborhood_boundaries/{0}/{1}/{2}{3}'.format(slugify(obj.organization.name),
+                                                           obj.state_abbrev or obj.country.alpha3,
+                                                           obj.name,
                                                            os.path.splitext(filename)[1])
 
 
@@ -138,7 +150,10 @@ class Neighborhood(PFBModel):
     organization = models.ForeignKey(Organization,
                                      related_name='neighborhoods',
                                      on_delete=models.CASCADE)
-    state_abbrev = USStateField(help_text='The US state of the uploaded neighborhood')
+    country = CountryField(default='US',
+                           help_text='The country of the uploaded neighborhood')
+    state_abbrev = USStateField(help_text='The state of the uploaded neighborhood, if in the US',
+                                blank=True, null=True)
     boundary_file = models.FileField(max_length=1024,
                                      upload_to=get_neighborhood_file_upload_path,
                                      help_text='A zipped shapefile boundary to run the ' +
@@ -151,7 +166,7 @@ class Neighborhood(PFBModel):
                                  on_delete=models.CASCADE, blank=True, null=True)
 
     def save(self, *args, **kwargs):
-        """ Override to do validation checks before saving, which disallows blank state_abbrev """
+        """ Override to do validation checks before saving """
         if not self.name:
             self.name = self.name_for_label(self.label)
         self.full_clean()
@@ -272,7 +287,8 @@ class Neighborhood(PFBModel):
                 shutil.rmtree(tmpdir, ignore_errors=True)
 
     class Meta:
-        unique_together = ('name', 'state_abbrev', 'organization',)
+        # Note that uniqueness fields should also be used in the upload file path
+        unique_together = ('name', 'country', 'state_abbrev', 'organization',)
 
 
 class AnalysisBatchManager(models.Manager):
@@ -491,7 +507,6 @@ class AnalysisJob(PFBModel):
 
     objects = AnalysisJobManager()
 
-
     @property
     def batch_job_status(self):
         """ Return current AWS Batch job status for this job
@@ -556,11 +571,25 @@ class AnalysisJob(PFBModel):
 
     @property
     def tile_urls(self):
-        return [{
-            'name': layer,
-            'url': self._s3_url_for_result_resource('tiles/neighborhood_{}'.format(layer) +
-                                                    '/{z}/{x}/{y}.png')
-        } for layer in ['ways', 'census_blocks', 'bike_infrastructure']]
+        layers = ['ways', 'census_blocks', 'bike_infrastructure']
+        tile_template = '{z}/{x}/{y}.png'
+        if hasattr(settings, 'TILEGARDEN_ROOT') and settings.TILEGARDEN_ROOT:
+            return [{
+                'name': layer,
+                'url': '{root}/tile/{job_id}/{layer}/{tile_template}'.format(
+                    root=settings.TILEGARDEN_ROOT,
+                    job_id=self.uuid,
+                    layer=layer,
+                    tile_template=tile_template)
+            } for layer in layers]
+        else:
+            # pre-Tilegarden URL format
+            return [{
+                'name': layer,
+                'url': self._s3_url_for_result_resource(
+                    'tiles/neighborhood_{layer}/{tile_template}'.format(
+                        layer=layer, tile_template=tile_template))
+            } for layer in layers]
 
     @property
     def overall_scores_url(self):
@@ -639,6 +668,13 @@ class AnalysisJob(PFBModel):
         """ Run the analysis job, configuring ENV appropriately """
         if self.status != self.Status.CREATED:
             logger.warn('Attempt to re-run job: {}. Skipping.'.format(self.uuid))
+            return
+
+        # TODO: #614 remove this check on adding support for running international jobs
+        if not self.neighborhood.state:
+            logger.warn('Running jobs outside the US is not supported yet. Skipping {}.'.format(
+                self.uuid))
+            self.update_status(self.Status.ERROR)
             return
 
         # Provide the base environment to enable runnin Django commands in the container
@@ -758,6 +794,41 @@ class AnalysisJob(PFBModel):
         )
 
 
+class NeighborhoodWaysResults(models.Model):
+    """ Stores geometries and results from the neighborhood ways shapefile.
+
+    """
+    uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    geom = LineStringField(srid=4326, blank=True, null=True)
+    job = models.ForeignKey(AnalysisJob,
+                            related_name='neighborhood_way_results',
+                            on_delete=models.CASCADE,
+                            null=True,
+                            blank=True)
+
+    tf_seg_str = models.PositiveSmallIntegerField(blank=True, null=True)
+    ft_seg_str = models.PositiveSmallIntegerField(blank=True, null=True)
+    xwalk = models.PositiveSmallIntegerField(blank=True, null=True)
+    ft_bike_in = models.CharField(blank=True, null=True, max_length=20)
+    tf_bike_in = models.CharField(blank=True, null=True, max_length=20)
+    functional = models.CharField(blank=True, null=True, max_length=20)
+
+
+class CensusBlocksResults(models.Model):
+    """ Stores geometries and results from the Census blocks shapefile.
+
+    """
+    uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    geom = MultiPolygonField(srid=4326, blank=True, null=True)
+    job = models.ForeignKey(AnalysisJob,
+                            related_name='census_block_results',
+                            on_delete=models.CASCADE,
+                            null=True,
+                            blank=True)
+
+    overall_score = models.FloatField(blank=True, null=True)
+
+
 class AnalysisJobStatusUpdate(models.Model):
     """ Related model for AnalysisJob, to provide record of status updates as job progresses
 
@@ -803,3 +874,27 @@ class AnalysisScoreMetadata(models.Model):
 
     class Meta:
         ordering = ('name',)
+
+
+class AnalysisLocalUploadTask(PFBModel):
+
+    class Status(object):
+        CREATED = 'created'
+        QUEUED = 'queued'
+        IMPORTING = 'importing'
+        COMPLETE = 'complete'
+        ERROR = 'error'
+
+        CHOICES = (
+            (CREATED, 'Created',),
+            (QUEUED, 'Queued',),
+            (IMPORTING, 'Importing',),
+            (COMPLETE, 'Complete',),
+            (ERROR, 'Error',),
+        )
+
+    status = models.CharField(max_length=16, choices=Status.CHOICES, default=Status.CREATED)
+    error = models.CharField(max_length=8192, blank=True, null=True)
+    job = models.OneToOneField(AnalysisJob, related_name='local_upload_task',
+                               on_delete=models.CASCADE)
+    upload_results_url = models.URLField(max_length=8192)