+++ src/django/pfb_network_connectivity/settings.py
@@ -73,6 +73,7 @@ INSTALLED_APPS = [
     'storages',
     'watchman',
     'django_q',
+    'django_countries',
 
     # Application
     'pfb_network_connectivity',
@@ -198,6 +199,10 @@ logging.config.dictConfig({
     },
     'loggers': {
         'django': {
+            'handlers': ['console'],
+            'level': DJANGO_LOG_LEVEL,
+        },
+        'django_q': {
             'handlers': ['console'],
             'level': DJANGO_LOG_LEVEL,
         },
@@ -340,3 +345,6 @@ PFB_ANALYSIS_DESTINATIONS = [
 ]
 # Length of time in seconds that S3 pre-signed urls are valid for
 PFB_ANALYSIS_PRESIGNED_URL_EXPIRES = 3600
+
+# Root URL for tile server. If unset, reverts to using uploaded S3 tile URLs
+TILEGARDEN_ROOT = os.getenv('PFB_TILEGARDEN_ROOT')